Refactored tests to re-utilize test scenarios between integrations tests projects.
