+++ tests/Rules.Framework.IntegrationTests.Common/Scenarios/Scenario1/ConditionTypes.cs
@@ -1,0 +1,7 @@
+namespace Rules.Framework.IntegrationTests.Common.Scenarios.Scenario1
+{
+    public enum ConditionTypes
+    {
+        Age = 1
+    }
+}
