+++ tests/Rules.Framework.IntegrationTests.Common/Scenarios/Scenario1/ContentTypes.cs
@@ -1,0 +1,7 @@
+namespace Rules.Framework.IntegrationTests.Common.Scenarios.Scenario1
+{
+    public enum ContentTypes
+    {
+        BodyMassIndexFormula = 1
+    }
+}
