+++ tests/Rules.Framework.IntegrationTests.Common/Scenarios/Scenario1/rules-framework-tests.body-mass-index.datasource.json
