+++ tests/Rules.Framework.IntegrationTests.Common/Scenarios/Scenario2/CarInsuranceAdvices.cs
@@ -1,0 +1,11 @@
+namespace Rules.Framework.IntegrationTests.Common.Scenarios.Scenario2
+{
+    public enum CarInsuranceAdvices
+    {
+        RefusePaymentPerFranchise = 1,
+
+        Pay = 2,
+
+        PayNewCar = 3
+    }
+}
