+++ tests/Rules.Framework.IntegrationTests.Common/Scenarios/Scenario2/ConditionTypes.cs
@@ -1,4 +1,4 @@
-namespace Rules.Framework.IntegrationTests.Tests.Scenario2
+namespace Rules.Framework.IntegrationTests.Common.Scenarios.Scenario2
 {
     public enum ConditionTypes
     {
