+++ tests/Rules.Framework.IntegrationTests.Common/Scenarios/Scenario2/rules-framework-tests.car-insurance-advisor.json
