--- tests/Rules.Framework.IntegrationTests/Tests/Scenario1/ConditionTypes.cs
@@ -1,7 +1,0 @@
-namespace Rules.Framework.IntegrationTests.Tests.Scenario1
-{
-    internal enum ConditionTypes
-    {
-        Age = 1
-    }
-}
