--- tests/Rules.Framework.IntegrationTests/Tests/Scenario1/ContentTypes.cs
@@ -1,7 +1,0 @@
-namespace Rules.Framework.IntegrationTests.Tests.Scenario1
-{
-    internal enum ContentTypes
-    {
-        BodyMassIndexFormula = 1
-    }
-}
