--- tests/Rules.Framework.IntegrationTests/Tests/Scenario2/CarInsuranceAdvices.cs
@@ -1,11 +1,0 @@
-namespace Rules.Framework.IntegrationTests.Tests.Scenario2
-{
-    internal enum CarInsuranceAdvices
-    {
-        RefusePaymentPerFranchise = 1,
-
-        Pay = 2,
-
-        PayNewCar = 3
-    }
-}
