--- tests/Rules.Framework.Providers.MongoDb.IntegrationTests/Tests/Scenario2/ConditionTypes.cs
@@ -1,11 +1,0 @@
-namespace Rules.Framework.Providers.MongoDb.IntegrationTests.Tests.Scenario2
-{
-    public enum ConditionTypes
-    {
-        RepairCosts = 1,
-
-        RepairCostsCommercialValueRate = 2,
-
-        SelfDamageCoverage = 3
-    }
-}
