--- tests/Rules.Framework.Providers.MongoDb.IntegrationTests/Tests/Scenario2/ContentTypes.cs
@@ -1,7 +1,0 @@
-namespace Rules.Framework.Providers.MongoDb.IntegrationTests.Tests.Scenario2
-{
-    public enum ContentTypes
-    {
-        CarInsuranceAdvice = 1
-    }
-}
