--- tests/Rules.Framework.Providers.MongoDb.IntegrationTests/Tests/Scenario3/SecuritySystemAction.cs
@@ -1,11 +1,0 @@
-namespace Rules.Framework.Providers.MongoDb.IntegrationTests.Tests.Scenario3
-{
-    using System;
-
-    internal class SecuritySystemAction
-    {
-        public Guid ActionId { get; set; }
-
-        public string? ActionName { get; set; }
-    }
-}
