--- tests/Rules.Framework.Providers.MongoDb.IntegrationTests/Tests/Scenario3/SecuritySystemActionables.cs
@@ -1,8 +1,0 @@
-namespace Rules.Framework.Providers.MongoDb.IntegrationTests.Tests.Scenario3
-{
-    public enum SecuritySystemActionables
-    {
-        FireSystem = 1,
-        PowerSystem = 2
-    }
-}
