--- tests/Rules.Framework.Providers.MongoDb.IntegrationTests/Tests/Scenario3/SecuritySystemConditions.cs
@@ -1,11 +1,0 @@
-namespace Rules.Framework.Providers.MongoDb.IntegrationTests.Tests.Scenario3
-{
-    public enum SecuritySystemConditions
-    {
-        TemperatureCelsius = 1,
-
-        SmokeRate = 2,
-
-        PowerStatus = 3
-    }
-}
